requirements.txt
streamlit
ultralytics
opencv-python
numpy
Pillow
streamlit-webrtc
aiortc

